+++ src/codec.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "compress")]
+pub mod bcj2;
 #[cfg(feature = "brotli")]
 pub mod brotli;
 #[cfg(feature = "lz4")]